Remove make from end-user workflow
