+++ config/samples/cluster.yaml
@@ -1,0 +1,18 @@
+apiVersion: "cluster.k8s.io/v1alpha1"
+kind: Cluster
+metadata:
+  name: test1
+spec:
+  clusterNetwork:
+    services:
+      cidrBlocks: ["10.96.0.0/12"]
+    pods:
+      cidrBlocks: ["192.168.0.0/16"]
+    serviceDomain: "cluster.local"
+  providerConfig:
+    value:
+      apiVersion: "awsprovider/v1alpha1"
+      kind: "AWSClusterProviderConfig"
+      region: "us-east-1"
+      sshKeyName: "default"
+
