+++ config/samples/machine-controlplane.yaml
@@ -1,0 +1,17 @@
+apiVersion: "cluster.k8s.io/v1alpha1"
+kind: Machine
+metadata:
+  generateName: aws-controlplane-
+  labels:
+    set: controlplane
+spec:
+  versions:
+    kubelet: v1.12.0
+    controlPlane: v1.12.0
+  providerConfig:
+    value:
+      apiVersion: awsprovider/v1alpha1
+      kind: AWSMachineProviderConfig
+      instanceType: "t2.medium"
+      iamInstanceProfile: "control-plane.cluster-api-provider-aws.sigs.k8s.io"
+      keyName: "default"
