+++ config/samples/machine-node.yaml
@@ -1,0 +1,16 @@
+apiVersion: "cluster.k8s.io/v1alpha1"
+kind: Machine
+metadata:
+  generateName: aws-node-
+  labels:
+    set: node
+spec:
+  versions:
+    kubelet: v1.12.0
+  providerConfig:
+    value:
+      apiVersion: awsprovider/v1alpha1
+      kind: AWSMachineProviderConfig
+      instanceType: "t2.medium"
+      iamInstanceProfile: "nodes.cluster-api-provider-aws.sigs.k8s.io"
+      keyName: "default"
